+++ packages/zefyr/lib/src/widgets/editable_text.dart
@@ -22,6 +22,29 @@ import 'selection.dart';
 import 'selection_controls.dart';
 import 'theme.dart';
 
+/// Operation mode of a Zefyr editable control.
+enum ZefyrMode {
+  /// Editing mode provides full access to all editing features: keyboard,
+  /// editor toolbar with formatting tools, selection controls and selection
+  /// toolbar with clipboard tools.
+  ///
+  /// Tapping on links in edit mode shows selection toolbar with contextual
+  /// actions instead of launching the link in a web browser.
+  edit,
+
+  /// Select-only mode allows users to select a range of text and have access
+  /// to selection toolbar including clipboard tools, contextual link tools
+  /// (open link, copy link) and any custom actions registered by
+  /// current [ZefyrSelectionControls] implementation.
+  ///
+  /// Tapping on links in select-only mode launches the link in a web browser.
+  select,
+
+  /// View-only mode disables almost all user interactions except the ability
+  /// to launch links in a web browser when tapped.
+  view,
+}
+
 /// Core widget responsible for editing Zefyr documents.
 ///
 /// Depends on presence of [ZefyrTheme] and [ZefyrScope] somewhere up the
@@ -36,7 +59,7 @@ class ZefyrEditableText extends StatefulWidget {
     @required this.focusNode,
     @required this.imageDelegate,
     this.autofocus: true,
-    this.enabled: true,
+    this.mode: ZefyrMode.edit,
     this.padding: const EdgeInsets.symmetric(horizontal: 16.0),
     this.physics,
     this.selectionControls,
@@ -46,7 +69,7 @@ class ZefyrEditableText extends StatefulWidget {
   final FocusNode focusNode;
   final ZefyrImageDelegate imageDelegate;
   final bool autofocus;
-  final bool enabled;
+  final ZefyrMode mode;
   final ScrollPhysics physics;
   final ZefyrSelectionControls selectionControls;
 
@@ -122,11 +145,11 @@ class _ZefyrEditableTextState extends State<ZefyrEditableText>
   void didUpdateWidget(ZefyrEditableText oldWidget) {
     super.didUpdateWidget(oldWidget);
     _updateSubscriptions(oldWidget);
-    if (!_didAutoFocus && widget.autofocus && widget.enabled) {
+    if (!_didAutoFocus && widget.autofocus && widget.mode == ZefyrMode.edit) {
       FocusScope.of(context).autofocus(focusNode);
       _didAutoFocus = true;
     }
-    if (!widget.enabled && focusNode.hasFocus) {
+    if (widget.mode != ZefyrMode.edit && focusNode.hasFocus) {
       _didAutoFocus = false;
       focusNode.unfocus();
     }
@@ -233,7 +256,7 @@ class _ZefyrEditableTextState extends State<ZefyrEditableText>
 
   // Triggered for both text and selection changes.
   void _handleLocalValueChange() {
-    if (widget.enabled &&
+    if (widget.mode == ZefyrMode.edit &&
         widget.controller.lastChangeSource == ChangeSource.local) {
       // Only request keyboard for user actions.
       requestKeyboard();